refactor(backups): _deltaVm → _incrementalVm
